+++ NEWS.md
@@ -1,3 +1,13 @@
+# lotri (development version)
+
+* New option of `cov` added which check for matrix suitability for
+  covariance matrix. When `cov=TRUE`, off-diagonal elements in
+  covariance matrices may no longer be nonzero if the diagonal value
+  is zero (rxode2#481). This will also check tht the matrix is
+  non-positive definite on the non-diagonal terms. `cov` can also be a
+  function to allow correction of the matrix to a positive definite
+  matrix automatically.
+
 # lotri 0.4.4
 
 * Bug fix for non-standard evaluation where you take the numeric vector from